move db and network call files to data package
