+++ app/src/main/java/com/droid/diexample/data/MainRepository.kt
@@ -1,22 +1,21 @@
-package com.droid.diexample.repository
+package com.droid.diexample.data
 
 import com.droid.diexample.model.Blog
-import com.droid.diexample.retrofit.BlogRetrofit
-import com.droid.diexample.retrofit.NetworkMapper
-import com.droid.diexample.room.BlogDao
-import com.droid.diexample.room.CacheMapper
+import com.droid.diexample.data.remote.api.BlogApiService
+import com.droid.diexample.data.remote.mapper.NetworkMapper
+import com.droid.diexample.data.local.dao.BlogDao
+import com.droid.diexample.data.local.mapper.CacheMapper
 import com.droid.diexample.util.DataState
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
-import java.security.PrivateKey
 
 class MainRepository
 constructor(
-        private val blogDao: BlogDao,
-        private val blogRetrofit: BlogRetrofit,
-        private val cacheMapper: CacheMapper,
-        private val networkMapper: NetworkMapper
+    private val blogDao: BlogDao,
+    private val blogRetrofit: BlogApiService,
+    private val cacheMapper: CacheMapper,
+    private val networkMapper: NetworkMapper
 ){
     suspend fun getBlogs(): Flow<DataState<List<Blog>>> = flow {
         emit(DataState.Loading)
