+++ app/src/main/java/com/droid/diexample/data/local/BlogDatabase.kt
@@ -1,7 +1,9 @@
-package com.droid.diexample.room
+package com.droid.diexample.data.local
 
 import androidx.room.Database
 import androidx.room.RoomDatabase
+import com.droid.diexample.data.local.entity.BlogCacheEntity
+import com.droid.diexample.data.local.dao.BlogDao
 
 @Database(entities = [BlogCacheEntity::class], version = 1)
 abstract class BlogDatabase : RoomDatabase(){
