+++ app/src/main/java/com/droid/diexample/data/local/entity/BlogCacheEntity.kt
@@ -1,4 +1,4 @@
-package com.droid.diexample.room
+package com.droid.diexample.data.local.entity
 
 import androidx.room.ColumnInfo
 import androidx.room.Entity
