+++ app/src/main/java/com/droid/diexample/di/RepositoryModule.kt
@@ -1,10 +1,10 @@
 package com.droid.diexample.di
 
-import com.droid.diexample.repository.MainRepository
-import com.droid.diexample.retrofit.BlogRetrofit
-import com.droid.diexample.retrofit.NetworkMapper
-import com.droid.diexample.room.BlogDao
-import com.droid.diexample.room.CacheMapper
+import com.droid.diexample.data.MainRepository
+import com.droid.diexample.data.remote.api.BlogApiService
+import com.droid.diexample.data.remote.mapper.NetworkMapper
+import com.droid.diexample.data.local.dao.BlogDao
+import com.droid.diexample.data.local.mapper.CacheMapper
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
@@ -18,10 +18,10 @@ object  RepositoryModule {
     @Provides
     fun provideMainRepository(
         blogDao: BlogDao,
-        retrofit: BlogRetrofit,
+        retrofit: BlogApiService,
         cacheMapper: CacheMapper,
         networkMapper: NetworkMapper
-    ) : MainRepository{
+    ) : MainRepository {
         return MainRepository(blogDao,retrofit,cacheMapper,networkMapper)
     }
 }
