+++ app/src/main/java/com/droid/diexample/di/RetrofitModule.kt
@@ -1,10 +1,6 @@
 package com.droid.diexample.di
 
-import com.droid.diexample.model.Blog
-import com.droid.diexample.retrofit.BlogNetworkEntity
-import com.droid.diexample.retrofit.BlogRetrofit
-import com.droid.diexample.retrofit.NetworkMapper
-import com.droid.diexample.util.EntityMapper
+import com.droid.diexample.data.remote.api.BlogApiService
 import com.google.gson.Gson
 import com.google.gson.GsonBuilder
 import dagger.Module
@@ -13,7 +9,6 @@ import dagger.hilt.InstallIn
 import dagger.hilt.android.components.ApplicationComponent
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
-import retrofit2.create
 import javax.inject.Singleton
 
 @Module
@@ -37,9 +32,9 @@ object RetrofitModule {
 
     @Singleton
     @Provides
-    fun  provideBlogService(retrofit: Retrofit.Builder): BlogRetrofit{
+    fun  provideBlogService(retrofit: Retrofit.Builder): BlogApiService {
         return retrofit
             .build()
-            .create(BlogRetrofit::class.java)
+            .create(BlogApiService::class.java)
     }
 }
