--- app/src/main/java/com/droid/diexample/retrofit/BlogRetrofit.kt
@@ -1,14 +1,0 @@
-package com.droid.diexample.retrofit
-
-import com.droid.diexample.model.Blog
-import com.droid.diexample.util.EntityMapper
-import retrofit2.http.GET
-
-interface BlogRetrofit {
-
-    @GET("blogs")
-    suspend fun get(): List<BlogNetworkEntity>
-}
-
-
-
